feat(video-comment): add DeleteType enum and change CommentTopType to int
Introduce DeleteType with UNDELETED, DELETED_BY_USER, and
DELETED_BY_VIDEO_CREATER for logical deletion. Change CommentTopType
value from short to int for consistency across the codebase.

+++ nilo-common/src/main/java/com/neon/nilocommon/entity/enums/videoComment/CommentTopType.java
@@ -9,5 +9,5 @@ public enum CommentTopType
 {
     NOT_TOP((short) 0), TOP((short) 1);
 
-    private final short value;
+    private final int value;
 }
